+++ apps/web/src/constants.ts
@@ -1,3 +1,4 @@
 export const DO_NOT_RENDER_ID_PREFIX = "do-not-render-";
 export const DOCS_LINK =
   "https://github.com/langchain-ai/open-agent-platform/blob/main/README.md";
+export const DESCRIPTION_MAX_LENGTH = 850;
